+++ zatgo_core/zatgo_core/report/zg_api_usage/zg_api_usage.js
@@ -1,3 +1,0 @@
-frappe.query_reports["ZG API Usage"] = {
-	filters: []
-};