+++ zatgo_core/zatgo_core/report/zg_audit_history/zg_audit_history.js
@@ -1,3 +1,0 @@
-frappe.query_reports["ZG Audit History"] = {
-	filters: []
-};