+++ zatgo_core/zatgo_core/report/zg_feature_flags/zg_feature_flags.js
@@ -1,3 +1,0 @@
-frappe.query_reports["ZG Feature Flags"] = {
-	filters: []
-};