+++ zatgo_core/zatgo_core/report/zg_installed_apps/zg_installed_apps.js
@@ -1,3 +1,0 @@
-frappe.query_reports["ZG Installed Apps"] = {
-	filters: []
-};