+++ zatgo_core/zatgo_core/report/zg_login_activity/zg_login_activity.js
@@ -1,3 +1,0 @@
-frappe.query_reports["ZG Login Activity"] = {
-	filters: []
-};