+++ zatgo_core/zatgo_core/report/zg_printer_usage/zg_printer_usage.js
@@ -1,3 +1,0 @@
-frappe.query_reports["ZG Printer Usage"] = {
-	filters: []
-};